managed/excelgptpro-azurefunction/requirements.txt
openai
azure-functions
requests
